+++ HISTORY.md
@@ -29,10 +29,12 @@
 - Made `TickTock.tick` void and `TickTock.tock` raise `TickTockException` if `.tick` not called first
 - Changed to `rapidjson` instead for built-in `json` module for all json operations for strictness and that sweet performance
 - Added `reverse_line_iterator` for iterating through a file backwards
+- Renamed `throws` to `raises`
 
 ### Bug fixes
 
 - Fixed a bug where a backslash would sometimes be printed from the logger before square brackets
+- Fixed `raises` throwing errors if error was not caught
 
 ## 0.6.7